gpu: nvgpu: common: fix MISRA Rule 10.4 Violations
MISRA Rule 10.4 only allows the usage of arithmetic operations on
operands of the same essential type category.

Adding "U" at the end of the integer literals or casting operands
to have same type of operands when an arithmetic operation is
performed.

This fixes violations where an arithmetic operation is performed on
signed and unsigned int types.

JIRA NVGPU-992

Reviewed-on: https://git-master.nvidia.com/r/1921459
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/gpu/nvgpu/common/ecc.c
@@ -169,7 +169,7 @@ int nvgpu_ecc_counter_init_per_fbpa(struct gk20a *g,
 	int num_fbpa = nvgpu_get_litter_value(g, GPU_LIT_NUM_FBPAS);
 	struct nvgpu_ecc_stat *stats;
 
-	stats = nvgpu_kzalloc(g, sizeof(*stats) * num_fbpa);
+	stats = nvgpu_kzalloc(g, sizeof(*stats) * (size_t)num_fbpa);
 	if (stats == NULL) {
 		return -ENOMEM;
 	}
